User Search Consolidation
-Consolidated teams.js slightly
-Created a universal-helpers.js file to handle any helper that is used in multiple files.
-Consolidated User searching among all search methods aside from Admin-Panel

-Admin Panel searches server-side to handle larger datasets
-All three other searches (Add Player, Roster, and Communities) is client side due to smaller datasets.

-Moved league functions to leagues.js from teams.js
-Removed redundant documentation and global imports
