+++ EsportsManagementTool/EsportsManagementTool/templates/dashboard.html
@@ -2406,6 +2406,7 @@
 <!-- Image Cropper -->
 <script src="https://cdnjs.cloudflare.com/ajax/libs/cropperjs/1.6.1/cropper.min.js"></script>
 <!-- Delete Modal js (must be called first -->
+<script src="{{ url_for('static', filename='universal-helpers.js') }}"></script>
 <script src="{{ url_for('static', filename='modalDelete.js') }}"></script>
 <script src="{{ url_for('static', filename='image-cropper.js') }}"></script>
 <!-- Rest of js imports -->
